Splited response object
